remove using physical BC value
This could get a little too restrictive for general cases. There might be cases
in 2D where there are not enough cells to correct fit a second order plane. A
better "fix" down the road may be to have a special interpolation routine for
setting the BC values.

Currently, this branch avoids using BC values that are not correctly
extrapolated.